+++ jmusic-player/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>jmusic-player</name>
+	<name>jmusicplayer</name>
 	<comment></comment>
 	<projects>
 	</projects>
