+++ jmusic-player/pom.xml
@@ -5,10 +5,10 @@
   <modelVersion>4.0.0</modelVersion>
 
   <groupId>com.ovalsquare</groupId>
-  <artifactId>jmusic-player</artifactId>
+  <artifactId>jmusicplayer</artifactId>
   <version>0.0.1-SNAPSHOT</version>
 
-  <name>jmusic-player</name>
+  <name>jmusicplayer</name>
   <!-- FIXME change it to the project's website -->
   <url>http://www.example.com</url>
 
@@ -16,6 +16,8 @@
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <maven.compiler.source>17</maven.compiler.source>
     <maven.compiler.target>17</maven.compiler.target>
+    <javafx.version>21</javafx.version>
+    <javafx.maven.plugin.version>0.0.8</javafx.maven.plugin.version>
   </properties>
   
   <build>
@@ -28,6 +30,13 @@
         <artifactId>maven-failsafe-plugin</artifactId>
         <version>2.22.2</version>
       </plugin>
+      
+    <plugin>
+        <groupId>org.openjfx</groupId>
+        <artifactId>javafx-maven-plugin</artifactId>
+        <version>0.0.8</version>
+    </plugin>
+
     </plugins>
   </build>
 
@@ -46,6 +55,12 @@
 	  <groupId>com.mpatric</groupId>
 	  <artifactId>mp3agic</artifactId>
 	  <version>0.9.1</version>
+	</dependency>
+	
+	<dependency>
+    	<groupId>org.openjfx</groupId>
+    	<artifactId>javafx-controls</artifactId>
+   	 	<version>17.0.0.1</version>
 	</dependency>
 
   </dependencies>
