+++ jmusic-player/src/main/java/com/ovalsquare/jmusicplayer/App.java
@@ -1,0 +1,28 @@
+package com.ovalsquare.jmusicplayer;
+
+import java.io.IOException;
+import java.util.ArrayList;
+
+import com.mpatric.mp3agic.InvalidDataException;
+import com.mpatric.mp3agic.UnsupportedTagException;
+
+/**
+ * Hello world!
+ *
+ */
+public class App 
+{
+
+    public static void main( String[] args ) throws UnsupportedTagException, InvalidDataException, IOException
+    {
+    	SongHandler sh = new SongHandler("A:\\");
+    	ArrayList<Song> songs = sh.createSongList();
+    	
+    	for (Song s : songs) {
+    		System.out.println(s.get_songPath());
+    	}
+    	
+  //  	GUI.main(args);
+
+    }
+}
