+++ jmusic-player/src/main/java/module-info.java
@@ -1,0 +1,8 @@
+module jmusicplayer {
+	requires javafx.controls;
+	requires mp3agic;
+	requires transitive javafx.graphics;
+	//requires junit;
+	
+	exports com.ovalsquare.jmusicplayer to javafx.graphics;
+}
