--- jmusic-player/src/test/java/com/ovalsquare/jmusic_player/AppTest.java
@@ -1,22 +1,0 @@
-package com.ovalsquare.jmusic_player;
-
-
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-
-import org.junit.jupiter.api.Test;
-
-/**
- * Unit test for simple App.
- */
-public class AppTest 
-{
-    /**
-     * Rigorous Test :-)
-     */
-    @Test
-    public void shouldAnswerWithTrue()
-    {
-        assertTrue( true );
-    }
-}
