+++ jmusic-player/src/test/java/com/ovalsquare/jmusicplayer/AppTest.java
@@ -1,0 +1,22 @@
+package com.ovalsquare.jmusicplayer;
+
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+
+import org.junit.jupiter.api.Test;
+
+/**
+ * Unit test for simple App.
+ */
+public class AppTest 
+{
+    /**
+     * Rigorous Test :-)
+     */
+    @Test
+    public void shouldAnswerWithTrue()
+    {
+        assertTrue( true );
+    }
+}
